updated terminal and user stats for combat

+++ modules/Terminal.mjs
@@ -1,4 +1,3 @@
-
 import Vec2 from "./Vec2.mjs";
 import ResourceManager from "./ResourceManager.mjs";
 import BaseWindow from "./BaseWindow.mjs";
@@ -33,6 +32,12 @@ class Terminal
     #fighting;
     #user;
     #enemy;
+    #sprite;
+    #bossName;
+    #bossSprite;
+    #deletables;
+    #safe;
+
 
     // HTML Elements
     #eWindow;
@@ -45,6 +50,9 @@ class Terminal
 
     // Data
     #fileSystem;
+    #file1;
+    #file2;
+    #file3;
 
 
 
@@ -69,10 +77,11 @@ class Terminal
         ["att", "cmdAttack"],
         ["inv", "cmdInventory"],
         ["inventory", "cmdInventory"],
-        ["use", "cmdUseItem"]
+        ["use", "cmdUseItem"],
+        ["reset", "cmdReset"]
     ]);
 
-    constructor(parent, id, filesys = new FileSystem(), directory = `C:\\>`, title = "C:\\Windows\\System32\\cmd.exe", icon = "terminal", position = new Vec2(450, 320), size = new Vec2(976, 512))
+    constructor(parent, id, filesys = new FileSystem(), file2 = new FileSystem(), file3 = new FileSystem(), directory = `C:\\>`, title = "C:\\Windows\\System32\\cmd.exe", icon = "terminal", position = new Vec2(450, 320), size = new Vec2(976, 512))
     {
         this.#baseWindow = new BaseWindow(parent, id, title, icon, position, size);
 
@@ -85,9 +94,14 @@ class Terminal
         this.#running = false;
         this.#fileSystem = filesys;
         this.#enCounter = 0;
-        this.#bossCounter = 2;
+        this.#bossCounter = 0;
+        this.#bossName = "Windows Defender"
+        this.#bossSprite = "images/ascii-images/enemies/msoft";
         this.#fighting = false;
-        this.#user = new User("Guest",20,20*Math.random(),Math.random(),Math.random());
+        this.#user = new User("Guest",20,5,Math.random(),Math.random());
+        this.#deletables = 36;
+        this.#file2 = file2;
+        this.#file3 = file3;
 
         this.#initialize();
     }
@@ -125,7 +139,7 @@ class Terminal
 
         // Bottom spacer block
         this.#eBlock = document.createElement("div");
-        this.#eBlock.setAttribute("class", "spacer");
+        this.#eBlock.setAttribute("class", "spacer"); 
         
         // The window content section
         this.#eTerminal = this.#baseWindow.getWindowElements().eContent;
@@ -136,17 +150,20 @@ class Terminal
         // The whole window
         this.#eWindow = this.#baseWindow.getWindowElements().eWindow;
 
-        this.#eTaskbar = document.getElementById("taskbar");
-        this.#eTaskbar.innerHTML = `<img class="taskbar-terminal-image" src="images/taskbar/taskbar-terminal.png" alt="terminal taskbar"><img class="taskbar-terminal-overlay" id="terminal-button"src="images/taskbar/taskbar-terminal.png" alt="terminal taskbar">`;
+        //this.#eTaskbar = document.getElementById("taskbar");
+        //this.#eTaskbar.innerHTML = `<img class="taskbar-terminal-image" src="images/taskbar/taskbar-terminal.png" alt="terminal taskbar"><img class="taskbar-terminal-overlay" id="terminal-button"src="images/taskbar/taskbar-terminal.png" alt="terminal taskbar">`;
         
+        this.#eWindow.setAttribute("class", "terminal window");
     }
 
     // Initializes cmd window event listeners
     #setupEventListeners()
     {
+        /*
+
         document.getElementById("terminal-button").addEventListener("click", () => {
             this.getBaseWindow().toggleMinimize();
-        });
+        });*/
         
         this.#eWindow.addEventListener("click", () => {
             this.#eActiveEntry.focus();
@@ -210,7 +227,7 @@ class Terminal
             this.printLine();
             this.#fileSystem.setLocation(args) ? this.#directory = args : this.printFile("resources/cmd-cd-error.txt");
             //Encounter chance
-            this.#enCounter += Math.random();
+            this.#enCounter += (Math.random() * 5);
         }
         this.awaitCommand();
     }
@@ -308,6 +325,8 @@ class Terminal
         }
         else if(this.#running)
         {
+            
+
             this.printLine();
             const file = this.#fileSystem.getFile(args);
             if (!file)
@@ -324,6 +343,8 @@ class Terminal
                 {
                     this.printLine("1 file key consumed!")
                     this.printLine("'" + args + "' was successfully unlocked and deleted")
+                    //Counting down files in map
+                    this.#deletables --;
                 }
                 else
                 {
@@ -332,8 +353,20 @@ class Terminal
             }
             else
             {
+                if(!this.#safe){
+                    //75% encounter chance when stealing files
+                    let fileFight = Math.random();
+                    if(fileFight <= .75){
+                    this.#fight(args, 10, "grunt", "images/ascii-images/enemies/file");
+                    this.awaitCommand();
+                    return;
+                    }
+                }
+
                 this.#fileSystem.deleteFile(args);
                 this.printLine("'" + args + "' was successfully deleted");
+                //Counting down files in map
+                this.#deletables --;
 
                 let chance = Math.random();
                 if (chance < 0.2)
@@ -349,6 +382,10 @@ class Terminal
                     {
                         this.printLine("1 file key added to inventory!");
                     }
+                }
+                if(this.#safe){
+                    this.#safe = false;
+                    return;
                 }
             }
             
@@ -398,10 +435,13 @@ class Terminal
 
     cmdInventory(args)
     {
-        this.printLine("---------- INVENTORY ----------");
-        this.printLine("corrupted files (healing): " + this.#user.inventory.items.get("heal"));
-        this.printLine("file keys (access): " + this.#user.inventory.items.get("keys"));
-        this.printLine("enter 'use [ ITEM_NAME ]' to use an item");
+        if(this.#running){
+            this.printLine("");
+            this.printLine("---------- INVENTORY ----------");
+            this.printLine("corrupted files (healing): " + this.#user.inventory.items.get("heal"));
+            this.printLine("file keys (access): " + this.#user.inventory.items.get("keys"));
+            this.printLine("enter 'use [ ITEM_NAME ]' to use an item");
+        }
         this.awaitCommand();
     }
 
@@ -457,23 +497,24 @@ class Terminal
             if (this.#user.removeItem("heal"))
             {
                 let healamt = this.#user.heal();
-                this.printLine("Corrupted File comnsumed successfully! " + healamt + " health points restored")
+                this.printLine("Corrupted File consumed successfully! " + healamt + " health points restored");
             }
             else
             {
-                this.printLine("ERROR: no corrupted files in inventory!")
+                this.printLine("ERROR: no corrupted files in inventory!");
             }
         }
 
         else if (args == "file key" || args == "filekey" || args == "key")
         {
-            this.printLine("ERROR: keys are used automatically when deleting locked files")
+            this.printLine("ERROR: keys are used automatically when deleting locked files");
         }
         
         else
         {
-            this.printLine("ERROR: no item '" + args + "' in inventory. Enter 'inv' to view inventory")
+            this.printLine("ERROR: no item '" + args + "' in inventory. Enter 'inv' to view inventory");
         }
+        this.awaitCommand();
     }
 
     #cmdError()
@@ -495,42 +536,37 @@ class Terminal
     ////////////////////////////////////////////////////////////////
     //  COMBAT COMMANDS
     ////////////////////////////////////////////////////////////////
-    #fight(){
-        //Bossmode
-        // if(this.#bossCounter >= 3){
-        //     this.#fighting = true;
-        //     let message = this.#fileSystem.getFileTree().getNode().name;
-        //     message += " boss attacks!!";
-        //     this.printLine(message);
-        //     this.#bossCounter = 0;
-        //     this.#user.boss();
-        // }
-        //Grunt (file)
-        if(this.#enCounter >= 1){
-            this.#baseWindow.setSize(new Vec2(700, 800));
-            this.#baseWindow.setPos(new Vec2(600, 120));
-            this.#user.grunt();
-            this.#fighting = true;
-            let message = this.#fileSystem.getDirectory().getName();
-            message += " attacks!";
-            this.#enemy = "images/ascii-images/enemies/file.txt";
-            this.printLine(message);
-            this.printLine("TYPE ATT TO ATTACK AND INV TO ACCESS INVENTORY");
-            this.#enCounter = 0;
-            this.#bossCounter += Math.random();
-            this.printCharacters(this.#enemy);
+
+    sleep(milliseconds) {
+        var start = new Date().getTime();
+        for (var i = 0; i < 1e7; i++) {
+          if ((new Date().getTime() - start) > milliseconds){
+            break;
+          }
         }
-        else if(this.#fighting){
-            let attack = this.#user.enemyAtt();
-            let message = this.#fileSystem.getDirectory().getName(); + " swings at you!";
-                this.printLine(message);
+      }
+
+
+      //Check counter, input enemyname, check type, return true if combat started false if not/going
+    #fight(enemyName, counter, enemyType, img){
+        if(this.#fighting){
+           
+            //this.sleep(10000)
+            let attack = this.#enemy.att(this.#user);
+            let message = this.#enemy.uName + " swings at you!";
+            this.printLine(message);
+            setTimeout(this.printCharacters(this.#sprite + ".txt"),5000);
+
             if(attack){
-                message = "Hits you for " + this.#user.damage + " damage!";
+                message = "Hits you for " + this.#enemy.damage + " damage!";
                 this.printLine(message);
                 if(this.#user.currentHp <= 0){
                     this.printLine("You've been defeated! Gameover!")
+
                     this.#fighting = false;
                     this.#running = false;
+                    this.#enCounter = 0;
+                    this.#bossCounter = 0;
                     this.cmdTerminalExe();
                     return;
                 }
@@ -539,48 +575,92 @@ class Terminal
                 message = "It misses!"
                 this.printLine(message);
             }
-            this.printCharacters(this.#enemy);
         }
+        else if(counter >= 10){
+            
+            this.#fighting = true;
+            if(enemyType == "boss"){
+                this.#baseWindow.setSize(new Vec2(900, 800));
+                this.#baseWindow.setPos(new Vec2(600, 120));
+                this.#enemy = new User(enemyName, Math.round(Math.random()*this.#user.hp*100)/100 + 4, 10, .5, .2);
+                this.printFile("resources/boss.txt")
+            }
+            else{
+                this.#baseWindow.setSize(new Vec2(800, 800));
+                this.#baseWindow.setPos(new Vec2(600, 120));
+                this.#enemy = new User(enemyName, Math.round(Math.random()*this.#user.hp*100)/100, 1 + Math.round(Math.random()*this.#user.str*100)/100, .25, .25);
+            }
+            this.printLine(enemyName + " attacks!!");
+            this.printLine("TYPE ATT TO ATTACK AND INV TO ACCESS INVENTORY");
+            this.#sprite = img;
+        }
+    
     }
 
-    printCharacters(enemy) 
+    printCharacters(sprite) 
     {   
-        this.printLine();
-        this.printLine();
-        let enemyStat = "                            ENEMY HP: " + this.#user.enemyChp + "/" + this.#user.enemyHp;
+        this.printLine("");
+        this.printLine("");
+        let enemyStat = "                            ENEMY HP: " + this.#enemy.currentHp + "/" + this.#enemy.hp;
         this.printLine(enemyStat);
-        this.printFile(enemy);
+        this.printFile(sprite);
         this.printFile("images/ascii-images/bacteriophage.txt");
-        let status = "HP: " + this.#user.currentHp + "/" + this.#user.hp;
-            this.printLine("USER STATUS:");
-            this.printLine(status);
-        this.printLine()
+        let status = "USER HP: " + this.#user.currentHp + "/" + this.#user.hp;
+        this.printLine(status);
+        this.printLine("")
+
     }
 
     cmdAttack(args){
         if(this.#fighting){
-            let attack = this.#user.att();
+            let attack = this.#user.att(this.#enemy);
             let message = "You attack!";
                 this.printLine(message);
+                
+            if(this.#enemy.currentHp <= 0){
+                this.#baseWindow.setSize(new Vec2(976, 512));
+                this.#baseWindow.setPos(new Vec2(450, 320));
+                this.#fighting = false;
+                
+                this.printCharacters(this.#sprite + "Dead.txt");
+                message = "You inflict " + this.#user.damage + " damage!";
+                this.printLine(message);
+                this.printLine("You defeated " + this.#enemy.uName + "!");
+                this.#user.exp(this.#enemy);
+                this.printLine("EXP gained: " + this.#enemy.str);
+                
+                if(this.#bossCounter >= 10){
+                    this.changePhase();
+                    this.#bossCounter = 0;
+                }
+                else if(this.#enCounter >= 10){
+                    this.#enCounter = 0;
+                }
+                else{
+                    this.#safe = true;
+                    this.cmdDelete(this.#enemy.uName);
+                }
+                this.$bossCounter += Math.random();
+                this.awaitCommand();
+                return;
+            }
+
             if(attack){
-                message = "You hit for " + this.#user.damage + " damage!";
+                if(this.#user.crit){
+                    this.printCharacters(this.#sprite + "Hot.txt");
+                    this.printLine("It's a critical hit!!")
+                }
+                else{
+                    this.printCharacters(this.#sprite + ".txt");
+                }
+                message = "You inflict " + this.#user.damage + " damage!";
                 this.printLine(message);
             }
             else{
                 message = "You miss!"
                 this.printLine(message);
-            }
-            if(this.#user.enemyChp <= 0){
-                this.#baseWindow.setSize(new Vec2(976, 512));
-                this.#baseWindow.setPos(new Vec2(450, 320));
-                this.#fighting = false;
-                this.printLine("You defeated " + this.#fileSystem.getDirectory().getName() + "!");
-                this.#user.exp();
-                this.printLine("EXP gained: " + this.#user.enemyStr);
-                this.awaitCommand();
-                return;
-            }
-            this.printCharacters(this.#enemy);
+            }           
+            
         }
         else if(!this.#running){
             this.printLine("Start the game first!");
@@ -589,7 +669,39 @@ class Terminal
             this.printLine("You lunge into the darkness!")
             this.printLine("...but there was nobody to attack.")
         }
-        this.awaitCommand()
+        this.awaitCommand();
+    }
+
+    changePhase(){
+        if(this.#bossName == "Windows Defender"){
+            this.printFile("phase1.txt");
+            this.#bossName = "Norton Antivirus";
+            this.#bossSprite = "images/ascii-images/enemies/norton";
+            this.#fileSystem = this.#file2;
+            this.#deletables = 36;
+        }
+        else if(this.#bossName == "Norton Antivirus"){
+            this.printFile("phase2.txt");
+            this.#bossName = "McAfee";
+            this.#bossSprite = "images/ascii-images/enemies/mcAfee";
+            this.#fileSystem = this.#file3;
+            this.#deletables = 36;
+
+        }
+        else{
+            this.printFile("phase3.txt");
+            this.#running = false;
+            this.printLine('TYPE "reset" TO PLAY AGAIN');
+
+        }
+    }
+    
+    cmdReset(){
+        this.#running = false;
+        this.#fileSystem = this.#file1;
+        //Reset stats
+        this.#user = new User("Guest",20,5,Math.random(),Math.random());            
+        this.cmdTerminalExe();
     }
 
     ////////////////////////////////////////////////////////////////
@@ -635,7 +747,19 @@ class Terminal
 
     awaitCommand()
     {
-        this.#fight();
+        if(!this.#fighting){
+            let filestuff = this.#fileSystem.getFileNames();
+            if(filestuff[0] != undefined){
+                this.#fight(filestuff[0], this.#enCounter, "grunt", "images/ascii-images/enemies/file");
+            }}
+        
+        this.#fight(this.#bossName, this.#bossCounter, "boss", this.#bossSprite);
+        if(this.#deletables == 0){
+            this.#bossCounter = 10;
+            this.#fight(this.#bossName, this.#bossCounter, "boss", this.#bossSprite);
+        }
+        
+        
         this.#awaitingCommand = true;
         this.printLine(this.#fileSystem.getPathString() + '>');
         this.enableInput();
